+++ src/Docker.Registry.DotNet/Endpoints/Implementations/SystemOperations.cs
@@ -1,0 +1,23 @@
+﻿using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+
+using Docker.Registry.DotNet.Registry;
+
+namespace Docker.Registry.DotNet.Endpoints.Implementations
+{
+    internal class SystemOperations : ISystemOperations
+    {
+        private readonly NetworkClient _client;
+
+        public SystemOperations(NetworkClient client)
+        {
+            this._client = client;
+        }
+
+        public Task PingAsync(CancellationToken cancellationToken = new CancellationToken())
+        {
+            return this._client.MakeRequestAsync(cancellationToken, HttpMethod.Get, "v2/");
+        }
+    }
+}
